+++ frontend-web/src/components/Header/index.js
@@ -1,10 +1,12 @@
 import React from 'react';
 
 import { Container } from './styles';
+import Icon from './Icon';
 
 function Header() {
   return (
     <Container>
+      <Icon />
       <span>{process.env.REACT_APP_NAME}</span>
     </Container>
   );