+++ frontend-web/src/pages/Dashboard/Constantes.js
@@ -1,8 +1,7 @@
 const emptyLancamento = {
-  os: undefined,
+  os: null,
   acao: '',
   intervalo: '',
-  sistema: undefined,
   copied: false,
   tarefaEvolutiva: false,
   isIntervalo: false,